brush: refactor moving each component to a its folder
