+++ brush/src/components/MapFooter/index.js
